magit-format-patch: support <since> commit
Do not convert a single commit argument read by magit-format-patch to
'<commit>^..<commit>' because this prevents the user from giving a
single commit argument that has the same meaning as it does for 'git
format-patch', where it is interpreted as '<commit>..HEAD'.  In place of
the conversion, try to provide better prompt defaults (including the
'<commit>^..<commit>' case).